+++ boden/include/windows-classic/bdn/ViewCore.h
@@ -33,7 +33,7 @@ public:
 	void setBounds(const Rect& bounds) override;
 
 
-	double uiLengthToPixels(const UiLength& uiLength) const override;
+	int uiLengthToPixels(const UiLength& uiLength) const override;
 	Margin uiMarginToPixelMargin(const UiMargin& margin) const override;
